+++ position/perft.hpp
@@ -2,7 +2,6 @@
 
 #include "movegen.hpp"
 #include <chrono>
-#include <iomanip>
 
 inline size_t perft(const int depth)
 {
@@ -10,13 +9,13 @@ inline size_t perft(const int depth)
         return 1;
     }
     const MoveList moves = legal_move_generator();
-    const size_t n = moves.size();
+    const int16_t n = moves.size();
     uint64_t nodes = 0;
     if (depth == 1) {
         return n;
     }
 
-    for (size_t i = 0; i < n; i++) {
+    for (int16_t i = 0; i < n; i++) {
         State st;
         position.make_move(moves.list[i], st);
         nodes += perft(depth - 1);
@@ -32,8 +31,8 @@ inline void divide(const int depth)
 
     const auto start = std::chrono::high_resolution_clock::now();
     const MoveList moves = legal_move_generator();
-    const size_t n = moves.size();
-    for (size_t i = 0; i < n; i++) {
+    const int16_t n = moves.size();
+    for (int16_t i = 0; i < n; i++) {
         State st;
         position.make_move(moves.list[i], st);
         const size_t num = perft(depth - 1);
@@ -44,5 +43,7 @@ inline void divide(const int depth)
     const auto time_taken = std::chrono::duration<double>(std::chrono::high_resolution_clock::now() - start);
     std::cout << "\nNodes searched: " << total << "\n";
     std::cout << "Time taken: " << std::chrono::duration_cast<std::chrono::milliseconds>(time_taken) << "\n";
-    std::cout << "Average: " << std::fixed << std::setprecision(2) << total / (std::chrono::duration_cast<std::chrono::nanoseconds>(time_taken).count() / 1000000000.0) << " nodes per second.\n\n";
+    std::cout << "Average: " << std::fixed << std::setprecision(2)
+              << total / (std::chrono::duration_cast<std::chrono::nanoseconds>(time_taken).count() /1000000000.0)
+              << " nodes per second.\n\n";
 }
